+++ tests/Traits/AdminMethodsEditTest.php
@@ -1,0 +1,80 @@
+<?php
+
+namespace Larrock\Core\Tests\Traits;
+
+use DaveJamesMiller\Breadcrumbs\BreadcrumbsServiceProvider;
+use Larrock\ComponentBlocks\BlocksComponent;
+use Larrock\ComponentBlocks\LarrockComponentBlocksServiceProvider;
+use Larrock\Core\LarrockCoreServiceProvider;
+use Larrock\Core\Tests\DatabaseTest\CreateBlocksDatabase;
+use Larrock\Core\Tests\DatabaseTest\CreateSeoDatabase;
+use Larrock\Core\Traits\AdminMethodsEdit;
+use Larrock\Core\Traits\ShareMethods;
+use Orchestra\Testbench\TestCase;
+use Proengsoft\JsValidation\JsValidationServiceProvider;
+
+class AdminMethodsEditTest extends TestCase
+{
+    protected function getEnvironmentSetUp($app)
+    {
+        $app['config']->set('database.default', 'sqlite');
+        $app['config']->set('database.connections.sqlite', [
+            'driver' => 'sqlite',
+            'database' => ':memory:',
+            'prefix' => '',
+        ]);
+    }
+
+    protected function setUp()
+    {
+        parent::setUp();
+
+        $seed = new CreateBlocksDatabase();
+        $seed->setUpBlocksDatabase();
+
+        $seed = new CreateSeoDatabase();
+        $seed->setUpSeoDatabase();
+    }
+
+    protected function getPackageProviders($app)
+    {
+        return [
+            LarrockCoreServiceProvider::class,
+            LarrockComponentBlocksServiceProvider::class,
+            BreadcrumbsServiceProvider::class,
+            JsValidationServiceProvider::class
+        ];
+    }
+
+    protected function getPackageAliases($app)
+    {
+        return [
+            'LarrockBlocks' => 'Larrock\ComponentBlocks\Facades\LarrockBlocks',
+            'Breadcrumbs' => 'DaveJamesMiller\Breadcrumbs\Facades\Breadcrumbs'
+        ];
+    }
+
+    public function testShareMethods()
+    {
+        $test = new AdminMethodsEditMock();
+        $this->assertCount(1, $test->shareMethods());
+    }
+
+    public function testEdit()
+    {
+        $test = new AdminMethodsEditMock();
+        $this->assertEquals('test', $test->edit(1)->getData()['data']->title);
+    }
+}
+
+class AdminMethodsEditMock
+{
+    use AdminMethodsEdit, ShareMethods;
+
+    protected $config;
+
+    public function __construct()
+    {
+        $this->config = new BlocksComponent();
+    }
+}